feat: Introduce BaseLLM and ILLMProvider for enhanced LLM abstraction
